fix(comment): profile and create profile component
